style: Split each signup field into separate components
