+++ frontend/app/types-constants/form.ts
@@ -1,0 +1,10 @@
+export type FormData = {
+  name: string;
+  screenName: string;
+  email: string;
+  password: string;
+};
+
+export const emailRegex =
+  // eslint-disable-next-line
+  /^[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;
